feat(anchor): add anchor to early access nav icon
